setup dependency injection
