+++ app/src/main/AndroidManifest.xml
@@ -8,9 +8,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
+        android:name=".Application"
         android:theme="@style/Theme.AndroidProjectTemplate">
         <activity
-            android:name=".ui.mainactivity.MainActivity"
+            android:name=".ui.mainActivity.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.AndroidProjectTemplate.NoActionBar">
             <intent-filter>
