+++ app/src/main/java/com/steve/mijiga/android/template/Application.kt
@@ -1,0 +1,7 @@
+package com.steve.mijiga.android.template
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class Application : Application()
