+++ app/src/main/java/com/steve/mijiga/android/template/di/AppModule.kt
@@ -1,0 +1,9 @@
+package com.steve.mijiga.android.template.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule
